feat: pagination and search
