--- .husky/pre-push
@@ -1,1 +1,0 @@
-npm run coverage
