--- src/App.tsx
@@ -1,36 +1,0 @@
-import { Routes, Route } from "react-router-dom";
-import MainPage from "./MainPage";
-import Page404 from "./app/not-found";
-import About from "./About";
-import ErrorBoundary from "./ErrorBoundary";
-import { BrowserRouter } from "react-router";
-import PlanetCard from "./PlanetCard";
-import { ThemeProvider } from "./ThemeProvider";
-import { Provider } from "react-redux";
-import { store } from "./store/store";
-import Wrapper from "./Wrapper";
-
-function App() {
-  return (
-    <ThemeProvider>
-      <Provider store={store}>
-        <ErrorBoundary>
-          <BrowserRouter>
-            <Routes>
-              <Route element={<Wrapper />}>
-                <Route path="/" element={<MainPage />} />
-                <Route path="/about" element={<About />} />
-                <Route path="/list/:pageNumber" element={<MainPage />}>
-                  <Route path=":planetId" element={<PlanetCard />} />
-                </Route>
-                <Route path="*" element={<Page404 />} />
-              </Route>
-            </Routes>
-          </BrowserRouter>
-        </ErrorBoundary>
-      </Provider>
-    </ThemeProvider>
-  );
-}
-
-export default App;
